+++ src/admin/pages/Post/Edit/PostFooter/Settings/index.tsx
