+++ src/admin/pages/Post/Edit/PostFooter/index.tsx
@@ -1,10 +1,10 @@
 import { Stack, Tooltip } from '@mui/material';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { LocalizedPost } from '../../../../types/index.js';
-import { IconButton } from '../../../@extended/components/IconButton/index.js';
-import { Icon } from '../../../components/elements/Icon/index.js';
-import { History } from './PostFooterContent/History/index.js';
+import { LocalizedPost } from '../../../../../types/index.js';
+import { IconButton } from '../../../../@extended/components/IconButton/index.js';
+import { Icon } from '../../../../components/elements/Icon/index.js';
+import { History } from './History/index.js';
 
 export type Props = {
   post: LocalizedPost;
