+++ src/admin/pages/Post/Edit/PostHeader/PublishSettings/index.tsx
@@ -32,7 +32,7 @@ import {
   editContentValidator,
 } from '../../../../../fields/validators/content/editContent.js';
 import { usePost } from '../../../Context/index.js';
-import AppBarStyled from '../../../PostHeader/AppBarStyled.js';
+import AppBarStyled from '../AppBarStyled.js';
 import { SlugSettings } from './SlugSettings/index.js';
 
 export type Props = {
